wip: Ceph + dokany bundle
We're considering an MSI bundle, wrapping Ceph and Dokany, and
maybe other dependencies as well in the future.

At the moment, it doesn't look too good. Each individual MSI
is displayed. If we're to use a custom template, we can hide the
individual MSI dialogs and provide an unified experience.